Refactor segment tracker into modular parts
