+++ lib/services/segment_tracker.dart
@@ -1,3 +1,5 @@
+library segment_tracker;
+
 import 'dart:async';
 import 'dart:convert';
 import 'dart:math' as math;
@@ -11,75 +13,14 @@ import 'package:toll_cam_finder/core/spatial/geo.dart';
 import 'package:toll_cam_finder/core/spatial/segment_geometry.dart';
 import 'package:toll_cam_finder/features/segemnt_index_service.dart';
 
-class SegmentTrackerEvent {
-  const SegmentTrackerEvent({
-    required this.startedSegment,
-    required this.endedSegment,
-    required this.activeSegmentId,
-    required this.debugData,
-  });
-
-  final bool startedSegment;
-  final bool endedSegment;
-  final String? activeSegmentId;
-  final SegmentTrackerDebugData debugData;
-}
-
-class SegmentTrackerDebugData {
-  const SegmentTrackerDebugData({
-    required this.isReady,
-    required this.querySquare,
-    required this.boundingCandidates,
-    required this.candidatePaths,
-    required this.startGeofenceRadius,
-    required this.endGeofenceRadius,
-  });
-
-  const SegmentTrackerDebugData.empty()
-      : isReady = false,
-        querySquare = const [],
-        boundingCandidates = const [],
-        candidatePaths = const [],
-        startGeofenceRadius = 0,
-        endGeofenceRadius = 0;
-
-  final bool isReady;
-  final List<LatLng> querySquare;
-  final List<SegmentGeometry> boundingCandidates;
-  final List<SegmentDebugPath> candidatePaths;
-  final double startGeofenceRadius;
-  final double endGeofenceRadius;
-
-  int get candidateCount => boundingCandidates.length;
-}
-
-class SegmentDebugPath {
-  const SegmentDebugPath({
-    required this.id,
-    required this.polyline,
-    required this.distanceMeters,
-    required this.isWithinTolerance,
-    required this.passesDirection,
-    required this.startHit,
-    required this.endHit,
-    required this.isActive,
-    required this.isDetailed,
-    this.nearestPoint,
-    this.headingDiffDeg,
-  });
-
-  final String id;
-  final List<LatLng> polyline;
-  final double distanceMeters;
-  final bool isWithinTolerance;
-  final bool passesDirection;
-  final bool startHit;
-  final bool endHit;
-  final bool isActive;
-  final bool isDetailed;
-  final LatLng? nearestPoint;
-  final double? headingDiffDeg;
-}
+part 'segment_tracker/segment_tracker_models.dart';
+part 'segment_tracker/segment_match.dart';
+part 'segment_tracker/active_segment_state.dart';
+part 'segment_tracker/segment_transition.dart';
+part 'segment_tracker/debug_extensions.dart';
+part 'segment_tracker/geometry_extensions.dart';
+part 'segment_tracker/heading_extensions.dart';
+part 'segment_tracker/path_fetch_extensions.dart';
 
 class SegmentTracker {
   SegmentTracker({
@@ -215,64 +156,6 @@ class SegmentTracker {
     }
   }
 
-  void _updateDebugData(
-    LatLng current,
-    List<SegmentGeometry> candidates,
-    List<SegmentMatch> matches,
-  ) {
-    final List<LatLng> square = _computeQuerySquare(current, candidateRadiusMeters);
-    final activeId = _active?.geometry.id;
-    final debugPaths = <SegmentDebugPath>[];
-
-    for (final match in matches) {
-      debugPaths.add(
-        SegmentDebugPath(
-          id: match.geometry.id,
-          polyline: List<LatLng>.unmodifiable(_pathToLatLngList(match.path)),
-          distanceMeters: match.distanceMeters,
-          isWithinTolerance: match.withinTolerance,
-          passesDirection: match.passesDirection,
-          startHit: match.startHit,
-          endHit: match.endHit,
-          isActive: match.geometry.id == activeId,
-          isDetailed: match.isDetailed,
-          nearestPoint: LatLng(match.nearestPoint.lat, match.nearestPoint.lon),
-          headingDiffDeg: match.headingDiffDeg,
-        ),
-      );
-    }
-
-    if (_active != null &&
-        _active!.lastMatch != null &&
-        !matches.any((m) => m.geometry.id == _active!.geometry.id)) {
-      final match = _active!.lastMatch!;
-      debugPaths.add(
-        SegmentDebugPath(
-          id: match.geometry.id,
-          polyline: List<LatLng>.unmodifiable(_pathToLatLngList(match.path)),
-          distanceMeters: match.distanceMeters,
-          isWithinTolerance: match.withinTolerance,
-          passesDirection: match.passesDirection,
-          startHit: match.startHit,
-          endHit: match.endHit,
-          isActive: true,
-          isDetailed: match.isDetailed,
-          nearestPoint: LatLng(match.nearestPoint.lat, match.nearestPoint.lon),
-          headingDiffDeg: match.headingDiffDeg,
-        ),
-      );
-    }
-
-    _latestDebugData = SegmentTrackerDebugData(
-      isReady: _isReady,
-      querySquare: List<LatLng>.unmodifiable(square),
-      boundingCandidates: List<SegmentGeometry>.unmodifiable(candidates),
-      candidatePaths: List<SegmentDebugPath>.unmodifiable(debugPaths),
-      startGeofenceRadius: startGeofenceRadiusMeters,
-      endGeofenceRadius: endGeofenceRadiusMeters,
-    );
-  }
-
   _SegmentTransition _updateActiveSegment(List<SegmentMatch> matches) {
     if (_active == null) {
       final entry = _chooseEntryMatch(matches);
@@ -374,131 +257,14 @@ class SegmentTracker {
     }
 
     if (!entry.isDetailed && fetchFailed) {
-      active.forceKeepUntilEnd = true;
-      active.enforceDirection = false;
+      active
+        ..forceKeepUntilEnd = true
+        ..enforceDirection = false;
     }
 
     if (kDebugMode) {
       debugPrint('[SEG] entered segment ${entry.geometry.id} '
           '(detailed=${entry.isDetailed}, fetchFailed=$fetchFailed)');
-    }
-  }
-
-  void _ensureEnhancedPath(SegmentGeometry geometry) {
-    if (geometry.path.length > 2) {
-      return;
-    }
-    if (_pathOverrides.containsKey(geometry.id)) {
-      return;
-    }
-    if (_fetchFailures.contains(geometry.id)) {
-      return;
-    }
-    if (_fetching.contains(geometry.id)) {
-      return;
-    }
-
-    _fetching.add(geometry.id);
-    unawaited(_downloadAndStorePath(geometry));
-  }
-
-  Future<void> _downloadAndStorePath(SegmentGeometry geometry) async {
-    List<GeoPoint>? enhanced;
-    try {
-      enhanced = await _fetchDetailedPath(geometry.path.first, geometry.path.last);
-    } finally {
-      if (enhanced != null && enhanced.length >= 2) {
-        final stored = List<GeoPoint>.unmodifiable(enhanced);
-        _pathOverrides[geometry.id] = stored;
-        _fetchFailures.remove(geometry.id);
-        if (_active != null && _active!.geometry.id == geometry.id) {
-          _active!
-            ..path = stored
-            ..forceKeepUntilEnd = false
-            ..enforceDirection = true
-            ..consecutiveMisses = 0
-            ..lastMatch = null;
-        }
-        if (kDebugMode) {
-          debugPrint('[SEG] enhanced path ready for ${geometry.id} (${enhanced.length} pts)');
-        }
-      } else {
-        _fetchFailures.add(geometry.id);
-        if (_active != null && _active!.geometry.id == geometry.id) {
-          _active!
-            ..forceKeepUntilEnd = true
-            ..enforceDirection = false;
-        }
-        if (kDebugMode) {
-          debugPrint('[SEG] enhanced path unavailable for ${geometry.id}');
-        }
-      }
-      _fetching.remove(geometry.id);
-    }
-  }
-
-  Future<List<GeoPoint>?> _fetchDetailedPath(GeoPoint start, GeoPoint end) async {
-    final uri = Uri.parse(
-      'https://router.project-osrm.org/route/v1/driving/'
-      '${start.lon},${start.lat};${end.lon},${end.lat}?overview=full&geometries=geojson',
-    );
-
-    try {
-      final response = await _httpClient.get(uri, headers: const {
-        'User-Agent': 'toll_cam_finder/segment-tracker',
-        'Accept': 'application/json',
-      });
-
-      if (response.statusCode != 200) {
-        if (kDebugMode) {
-          debugPrint('[SEG] enhanced path status ${response.statusCode} for ${uri.path}');
-        }
-        return null;
-      }
-
-      final dynamic decoded = json.decode(response.body);
-      if (decoded is! Map<String, dynamic>) return null;
-
-      final routes = decoded['routes'];
-      if (routes is! List || routes.isEmpty) return null;
-      final route = routes.first;
-      if (route is! Map<String, dynamic>) return null;
-
-      dynamic geometry = route['geometry'];
-      List? coords;
-      if (geometry is Map<String, dynamic> && geometry['coordinates'] is List) {
-        coords = geometry['coordinates'] as List;
-      } else if (geometry is List) {
-        coords = geometry;
-      }
-      if (coords == null) return null;
-
-      final path = <GeoPoint>[];
-      for (final coord in coords) {
-        if (coord is List && coord.length >= 2) {
-          final lon = (coord[0] as num).toDouble();
-          final lat = (coord[1] as num).toDouble();
-          path.add(GeoPoint(lat, lon));
-        }
-      }
-
-      if (path.length < 2) {
-        return null;
-      }
-
-      if (_distanceBetween(path.first, start) > 5) {
-        path.insert(0, start);
-      }
-      if (_distanceBetween(path.last, end) > 5) {
-        path.add(end);
-      }
-
-      return path;
-    } catch (e) {
-      if (kDebugMode) {
-        debugPrint('[SEG] failed to fetch enhanced path: $e');
-      }
-      return null;
     }
   }
 
@@ -508,241 +274,4 @@ class SegmentTracker {
     }
     _active = null;
   }
-
-  List<GeoPoint> _effectivePathFor(SegmentGeometry geom) {
-    return _pathOverrides[geom.id] ?? geom.path;
-  }
-
-  bool _isPathDetailed(SegmentGeometry geom, List<GeoPoint> path) {
-    if (_pathOverrides.containsKey(geom.id)) {
-      return true;
-    }
-    return path.length > 2;
-  }
-
-  List<LatLng> _pathToLatLngList(List<GeoPoint> path) {
-    return path.map((p) => LatLng(p.lat, p.lon)).toList(growable: false);
-  }
-
-  List<LatLng> _computeQuerySquare(LatLng center, double radiusMeters) {
-    const double mPerDegLat = 111320.0;
-    final double latRad = center.latitude * math.pi / 180.0;
-    final double mPerDegLon = (mPerDegLat * math.cos(latRad)).clamp(1e-9, double.infinity);
-    final double dLat = radiusMeters / mPerDegLat;
-    final double dLon = radiusMeters / mPerDegLon;
-
-    final double minLat = center.latitude - dLat;
-    final double maxLat = center.latitude + dLat;
-    final double minLon = center.longitude - dLon;
-    final double maxLon = center.longitude + dLon;
-
-    return <LatLng>[
-      LatLng(minLat, minLon),
-      LatLng(minLat, maxLon),
-      LatLng(maxLat, maxLon),
-      LatLng(maxLat, minLon),
-      LatLng(minLat, minLon),
-    ];
-  }
-
-  double? _resolveHeading(
-    LatLng current,
-    LatLng? previous,
-    double? rawHeading,
-    double? speedKmh,
-  ) {
-    double? heading = _normalizeHeading(rawHeading);
-    final bool speedOk = speedKmh == null || speedKmh >= directionMinSpeedKmh;
-    if (heading != null && speedOk) {
-      return heading;
-    }
-
-    if (previous != null) {
-      final GeoPoint prevPoint = GeoPoint(previous.latitude, previous.longitude);
-      final GeoPoint currPoint = GeoPoint(current.latitude, current.longitude);
-      if (_distanceBetween(prevPoint, currPoint) >= 3) {
-        return _bearingBetween(previous, current);
-      }
-    }
-
-    return heading;
-  }
-
-  double? _normalizeHeading(double? heading) {
-    if (heading == null || !heading.isFinite) return null;
-    double value = heading % 360.0;
-    if (value < 0) value += 360.0;
-    return value;
-  }
-
-  double _bearingBetween(LatLng from, LatLng to) {
-    final double lat1 = from.latitude * math.pi / 180.0;
-    final double lat2 = to.latitude * math.pi / 180.0;
-    final double dLon = (to.longitude - from.longitude) * math.pi / 180.0;
-    final double y = math.sin(dLon) * math.cos(lat2);
-    final double x = math.cos(lat1) * math.sin(lat2) -
-        math.sin(lat1) * math.cos(lat2) * math.cos(dLon);
-    double bearing = math.atan2(y, x) * 180.0 / math.pi;
-    if (bearing < 0) bearing += 360.0;
-    return bearing;
-  }
-
-  double _angularDifferenceDegrees(double a, double b) {
-    double diff = (a - b).abs() % 360.0;
-    if (diff > 180.0) diff = 360.0 - diff;
-    return diff;
-  }
-
-  double _distanceBetween(GeoPoint a, GeoPoint b) {
-    const double earthRadius = 6371000.0; // meters
-    final double lat1 = a.lat * math.pi / 180.0;
-    final double lat2 = b.lat * math.pi / 180.0;
-    final double dLat = lat2 - lat1;
-    final double dLon = (b.lon - a.lon) * math.pi / 180.0;
-
-    final double sinLat = math.sin(dLat / 2);
-    final double sinLon = math.sin(dLon / 2);
-    final double h = sinLat * sinLat + math.cos(lat1) * math.cos(lat2) * sinLon * sinLon;
-    final double c = 2 * math.atan2(math.sqrt(h), math.sqrt(1 - h));
-    return earthRadius * c;
-  }
-
-  _NearestPointResult _nearestPointOnPath(GeoPoint point, List<GeoPoint> path) {
-    const double mPerDegLat = 111320.0;
-    final double latRad = point.lat * math.pi / 180.0;
-    final double mPerDegLon = (mPerDegLat * math.cos(latRad)).clamp(1e-9, double.infinity);
-    final double refLat = point.lat;
-    final double refLon = point.lon;
-
-    double bestDistSq = double.infinity;
-    double bestX = 0;
-    double bestY = 0;
-    double? bestBearingDeg;
-
-    final offsets = <math.Point<double>>[];
-    for (final p in path) {
-      final double x = (p.lon - refLon) * mPerDegLon;
-      final double y = (p.lat - refLat) * mPerDegLat;
-      offsets.add(math.Point<double>(x, y));
-    }
-
-    for (var i = 0; i < offsets.length - 1; i++) {
-      final math.Point<double> a = offsets[i];
-      final math.Point<double> b = offsets[i + 1];
-      final double dx = b.x - a.x;
-      final double dy = b.y - a.y;
-      final double lenSq = dx * dx + dy * dy;
-      if (lenSq <= 1e-6) continue;
-
-      var t = (-(a.x * dx + a.y * dy)) / lenSq;
-      if (t < 0) t = 0;
-      if (t > 1) t = 1;
-
-      final double projX = a.x + dx * t;
-      final double projY = a.y + dy * t;
-      final double distSq = projX * projX + projY * projY;
-      if (distSq < bestDistSq) {
-        bestDistSq = distSq;
-        bestX = projX;
-        bestY = projY;
-        final double bearingRad = math.atan2(dx, dy);
-        double bearingDeg = bearingRad * 180.0 / math.pi;
-        if (bearingDeg < 0) bearingDeg += 360.0;
-        bestBearingDeg = bearingDeg;
-      }
-    }
-
-    if (bestDistSq.isInfinite && offsets.isNotEmpty) {
-      final math.Point<double> first = offsets.first;
-      bestX = first.x;
-      bestY = first.y;
-      bestDistSq = bestX * bestX + bestY * bestY;
-      if (offsets.length >= 2) {
-        final math.Point<double> next = offsets[1];
-        final double dx = next.x - first.x;
-        final double dy = next.y - first.y;
-        final double bearingRad = math.atan2(dx, dy);
-        double bearingDeg = bearingRad * 180.0 / math.pi;
-        if (bearingDeg < 0) bearingDeg += 360.0;
-        bestBearingDeg = bearingDeg;
-      }
-    }
-
-    final double nearestLat = refLat + (bestY / mPerDegLat);
-    final double nearestLon = refLon + (bestX / mPerDegLon);
-    final double distance = math.sqrt(bestDistSq.abs());
-
-    return _NearestPointResult(
-      distanceMeters: distance,
-      point: GeoPoint(nearestLat, nearestLon),
-      bearingDeg: bestBearingDeg,
-    );
-  }
-}
-
-class SegmentMatch {
-  SegmentMatch({
-    required this.geometry,
-    required this.path,
-    required this.distanceMeters,
-    required this.nearestPoint,
-    required this.startDistanceMeters,
-    required this.endDistanceMeters,
-    required this.headingDiffDeg,
-    required this.withinTolerance,
-    required this.startHit,
-    required this.endHit,
-    required this.passesDirection,
-    required this.isDetailed,
-  });
-
-  final SegmentGeometry geometry;
-  final List<GeoPoint> path;
-  final double distanceMeters;
-  final GeoPoint nearestPoint;
-  final double startDistanceMeters;
-  final double endDistanceMeters;
-  final double? headingDiffDeg;
-  final bool withinTolerance;
-  final bool startHit;
-  final bool endHit;
-  final bool passesDirection;
-  final bool isDetailed;
-}
-
-class _ActiveSegmentState {
-  _ActiveSegmentState({
-    required this.geometry,
-    required this.path,
-    required this.forceKeepUntilEnd,
-    required this.enforceDirection,
-    required this.enteredAt,
-  });
-
-  final SegmentGeometry geometry;
-  List<GeoPoint> path;
-  bool forceKeepUntilEnd;
-  bool enforceDirection;
-  final DateTime enteredAt;
-  int consecutiveMisses = 0;
-  SegmentMatch? lastMatch;
-}
-
-class _NearestPointResult {
-  const _NearestPointResult({
-    required this.distanceMeters,
-    required this.point,
-    this.bearingDeg,
-  });
-
-  final double distanceMeters;
-  final GeoPoint point;
-  final double? bearingDeg;
-}
-
-class _SegmentTransition {
-  const _SegmentTransition({this.started = false, this.ended = false});
-
-  final bool started;
-  final bool ended;
 }
