+++ lib/services/segment_tracker/segment_transition.dart
@@ -1,0 +1,8 @@
+part of segment_tracker;
+
+class _SegmentTransition {
+  const _SegmentTransition({this.started = false, this.ended = false});
+
+  final bool started;
+  final bool ended;
+}
